requirements.txt
flask
flask-cors
google-cloud-storage
google-cloud-firestore
python-dotenv
werkzeug
pyjwt
firebase-admin
gunicorn